add linear diffusion
