+++ Project.toml
@@ -25,7 +25,7 @@ Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 CUDA = "5"
 DiffEqBase = "6"
 DiffEqCallbacks = "3, 4"
-GeoParams = "0.7.6"
+GeoParams = "0.7.7"
 HDF5 = "0.17"
 MuladdMacro = "0.2.4"
 OrdinaryDiffEqStabilizedRK = "1"
