+++ src/DiffusionGarnet.jl
@@ -87,7 +87,9 @@ include("callbacks/output_paraview.jl")
 include("discretisation/1D/semi_discretisation_1D_major.jl")
 include("discretisation/1D/semi_discretisation_1D_trace.jl")
 include("discretisation/2D/semi_discretisation_2D_major.jl")
+include("discretisation/2D/semi_discretisation_2D_trace.jl")
 include("discretisation/3D/semi_discretisation_3D_major.jl")
+include("discretisation/3D/semi_discretisation_3D_trace.jl")
 include("discretisation/spherical/semi_discretisation_spherical_major.jl")
 include("discretisation/spherical/semi_discretisation_spherical_trace.jl")
 include("simulate/simulate.jl")
